[Refactor] 전체 로직 정리

+++ src/main/java/groom/him/domain/notice/models/dto/response/NoticeResponse.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 public record NoticeResponse(
     Long noticeId,
-    Long memberId,
+    Integer memberId,
     String title,
     String content,
     IsPublic isPublic,
